refactor(logger): extract redact() into scripts/redact.js (activity-monitor)
The .frame/bin/ scripts run in their own processes and cannot reach
src/main/, so the shared patterns move to a dependency-free module both
sides require. logger.js re-exports them; its public API is unchanged.
